Improve `:Compile` so it works more like the Emacs counterpart.
Previously this command splits all the user-provided args by spaces, and
if you want a argument that contains spaces, you'll need to backslash
quote the spaces, which is very annoying.

For example, previously if you would like to do `echo hello && sleep 1`,
this is not possible, because everything will be passed to the
subprocess literally, i. e. `echo "hello" "&&" "sleep" "1"`.

Now that this command takes all the user input as a whole string, and
you can do whatever you want in the arguments, which is more convenient
and robust, and this style matches more like the `M-x compile` we can
find in Emacs.